v6.7.8 - add Strategies.Header.horizontalScroll to dist, add mutation testing, MUI DataGrid Presets refactor
